Change worker to record each incoming tweet
Re-jiggered the worker process to start recording each tweet as it comes in.
This could bite me in the ass, but it's worth a shot if we can swing the
overhead associated with storing that much data.

Started from scratch in the API to match the new data architecture. Now
have the following endpoints:
* /api/meta/start_date
* /api/count
* /api/count/since/:start
* /api/count/until/:end
* /api/count/from/:from/to/:to